introduce PublisherUtils.py and use in TaskPublishRucio.py
